Fix all Bug, Make map 1_4

+++ 05-ScenceManager/Floor.cpp
@@ -1,6 +1,6 @@
 #include "Floor.h"
 
-CFloor::CFloor(int width, int height)
+CFloor::CFloor(float width, int height)
 {
 	this->width = width;
 	this->height = height;
